rule evaluation implementation-first 4 rules

+++ src/main/java/com/edstem/project/contract/response/ProductResponse.java
@@ -10,5 +10,5 @@ import lombok.*;
 @Builder
 public class ProductResponse {
     private String productId;
-    private Integer stockLevel;
+    private Double stockLevel;
 }
